Fix nasty bug in ICall framework, inc. heap
Fixed a nasty bug in the ICall framework on line 1241 of ICall.c wherein
it was trying to start a timer with timeout 0 (not allowed) and getting
stuck in an ON error loop. This fix doesn't fix whatever was causing this,
but afterwards it all seems to work.

Increased the size of the heap because we were running out. Commented out
FSM.c. We don't have enough memory for that function pointer struct
unfortunately.