Opt. memory. Restructure pmanager code optimally.
Optimized memory by reducing extra semaphores, reduce I2C stack size (it
doesn't change much), and set the compiler to optimization level 5.

Restructured the peripheral manager code so that it configures each device
and starts its conversion cycle on each cycle, which is more akin to what
we'll see with proper power management. This method also optimally
interweaves the conversion time for all of the devices with other
operations.

Closes #19 #17 #7